safari-534.53-branch/LayoutTests/http/tests/local/blob/send-data-blob-expected.txt

CONSOLE MESSAGE: line 134: ReferenceError: Can't find variable: WebKitBlobBuilder
Test for building blobs with multiple strings combined by BlobBuilder and sending them via XMLHttpRequest.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* BlobBuilder.append(string)
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
